[CD-7541]: Launch Template support in AMI Deployment (#10712)
* [CD-7541]: added support for creating asg using base launch template

* [CD-7541]: updated logic to fetch launch template from Autoscaling group.

* [CD-7541]: updated logic to fetch launch template from Autoscaling group.

* [CD-7541]: added some unit tests

* [CD-7541]: added some more unit tests

* [CD-7541]: Renamed oldLaunchTemplateVersion fields in AwsAmiHelperServiceDelegateImpl.java to baseLaunchTemplateVersion

+++ 71-rest/src/main/java/software/wings/service/impl/aws/delegate/AwsAsgHelperServiceDelegateImpl.java
@@ -290,15 +290,17 @@ public class AwsAsgHelperServiceDelegateImpl
               amazonAutoScalingClient, autoScalingGroup.getAutoScalingGroupName(), new HashSet<>(), callback, true);
           logger.warn("Failed to delete ASG: [{}] [{}]", autoScalingGroup.getAutoScalingGroupName(), ignored);
         }
-        try {
-          tracker.trackASGCall("Delete Launch Config");
-          amazonAutoScalingClient.deleteLaunchConfiguration(
-              new DeleteLaunchConfigurationRequest().withLaunchConfigurationName(
-                  autoScalingGroup.getLaunchConfigurationName()));
-        } catch (Exception ignored) {
-          describeAutoScalingGroupActivities(
-              amazonAutoScalingClient, autoScalingGroup.getAutoScalingGroupName(), new HashSet<>(), callback, true);
-          logger.warn("Failed to delete ASG: [{}] [{}]", autoScalingGroup.getAutoScalingGroupName(), ignored);
+        if (isNotEmpty(autoScalingGroup.getLaunchConfigurationName())) {
+          try {
+            tracker.trackASGCall("Delete Launch Config");
+            amazonAutoScalingClient.deleteLaunchConfiguration(
+                new DeleteLaunchConfigurationRequest().withLaunchConfigurationName(
+                    autoScalingGroup.getLaunchConfigurationName()));
+          } catch (Exception ignored) {
+            describeAutoScalingGroupActivities(
+                amazonAutoScalingClient, autoScalingGroup.getAutoScalingGroupName(), new HashSet<>(), callback, true);
+            logger.warn("Failed to delete ASG: [{}] [{}]", autoScalingGroup.getAutoScalingGroupName(), ignored);
+          }
         }
       });
     } catch (AmazonServiceException amazonServiceException) {
